relay-server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
websockets==14.2
loguru==0.7.3
aiohttp==3.11.12
